6 unit test

+++ src/cmd/skip.ts
@@ -2,7 +2,7 @@ import { ChatInputCommandInteraction, SlashCommandBuilder } from 'discord.js';
 import { getVoiceConnection } from '@discordjs/voice';
 import ClientError from '../clientError';
 import GuildPlayer from '../guildPlayer';
-import { t } from '../i18next';
+import i18next from 'i18next';
 import PlayerService from '../playerService';
 
 export default {
@@ -12,18 +12,22 @@ export default {
 
 	async execute(interaction: ChatInputCommandInteraction<'cached'>): Promise<void> {
 		if (!getVoiceConnection(interaction.guildId)) {
-			throw new ClientError(t('notInServer'));
+			throw new ClientError(i18next.t('errors.music.notInServer'));
 		}
+		if (!interaction.channel || !interaction.channel.isTextBased()) {
+			throw new ClientError(i18next.t('errors.cmd.commandInTextChannel'));
+		}
+
 		const playerService: PlayerService = PlayerService.getInstance();
 		const player: GuildPlayer = playerService.getGuildPlayer(interaction.guildId);
 		const track_name: string | undefined = player.skip();
 		if (track_name) {
-			player.updateChannelId(interaction.channelId, interaction.channel);
-			await interaction.reply(t('skipTrack', { trackName: track_name }));
+			player.updateChannelId(interaction.channelId);
+			await interaction.reply(i18next.t('music.skipTrack', { trackName: track_name }));
 		}
 		else {
 			playerService.deleteGuildPlayer(interaction.guildId);
-			await interaction.reply(t('stopNoTracks'));
+			await interaction.reply(i18next.t('music.stopNoTracks'));
 		}
 	},
 };
